Use golangci-lint[1] and the depguard[2] linter to enforce that 'io/ioutil' is not used anymore
Instead of using individual linters, we rely on golangci-lint,
which is a Go linters aggregrator that runs linters in parallel.

NOTE: ideally, we should use all linters enabled by default by golangci-lint.
But doing so reports a lot of issues, not relevant to the issue here
(which is to enforce that 'io/ioutil' is not used anymore).
Enabling the default linters and fixing the issues reported
will be done in a subsequent issue/PR.

[1] https://golangci-lint.run/
[2] https://golangci-lint.run/usage/linters/#depguard

Signed-off-by: Armel Soro <asoro@redhat.com>