use zap.NewDevelopmentConfig to setup logger
The Zap library provides a logger config that can
be used without pulling in controller-runtime as a
dependency. Since alizer is only using a few options,
the dependency on controller-runtime can be avoided
by using zap's NewDevelopmentConfig (which sets
the development flag and the ISO8601 time encoder),
and setting the log level on that.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>